Refactor: Put transformations in their own files to make them easier to find
