+++ lib/inline_svg/transform_pipeline/transformations/class_attribute.rb
@@ -1,0 +1,10 @@
+module InlineSvg::TransformPipeline::Transformations
+  class ClassAttribute < Transformation
+    def transform(doc)
+      doc = Nokogiri::XML::Document.parse(doc.to_html)
+      svg = doc.at_css 'svg'
+      svg['class'] = value
+      doc
+    end
+  end
+end
