+++ lib/inline_svg/transform_pipeline/transformations/no_comment.rb
@@ -1,0 +1,10 @@
+module InlineSvg::TransformPipeline
+  module Transformations
+    class NoComment < Transformation
+      def transform(doc)
+        doc = Loofah::HTML::DocumentFragment.parse(doc.to_html)
+        doc.scrub!(:strip)
+      end
+    end
+  end
+end
